consolidate db.tables module
