+++ mathesar/rpc/tables/base.py
@@ -6,10 +6,15 @@ from typing import Literal, Optional, TypedDict
 from modernrpc.core import rpc_method, REQUEST_KEY
 from modernrpc.auth.basic import http_basic_auth_login_required
 
-from db.tables.operations.select import get_table_info, get_table, list_joinable_tables, get_preview
-from db.tables.operations.drop import drop_table_from_database
-from db.tables.operations.create import create_table_on_database
-from db.tables.operations.alter import alter_table_on_database
+from db.tables import (
+    alter_table_on_database,
+    create_table_on_database,
+    drop_table_from_database,
+    get_preview,
+    get_table,
+    get_table_info,
+    list_joinable_tables,
+)
 from mathesar.imports.csv import import_csv
 from mathesar.rpc.columns import CreatableColumnInfo, SettableColumnInfo, PreviewableColumnInfo
 from mathesar.rpc.constraints import CreatableConstraintInfo
